+++ example/windows/flutter/generated_plugins.cmake
@@ -6,6 +6,7 @@ list(APPEND FLUTTER_PLUGIN_LIST
   bitsdojo_window_windows
   flutter_acrylic
   system_theme
+  url_launcher_windows
 )
 
 set(PLUGIN_BUNDLED_LIBRARIES)