+++ lib/src/app.dart
@@ -1,5 +1,4 @@
 import 'package:fluent_ui/fluent_ui.dart';
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart' show DefaultMaterialLocalizations;
 
 /// An application that uses fluent design.