+++ lib/src/controls/inputs/buttons/base.dart
@@ -3,8 +3,6 @@ import 'package:flutter/foundation.dart';
 
 import 'package:fluent_ui/fluent_ui.dart';
 
-import 'theme.dart';
-
 abstract class BaseButton extends StatefulWidget {
   const BaseButton({
     Key? key,