+++ lib/src/controls/inputs/buttons/icon_button.dart
@@ -1,8 +1,5 @@
 import 'package:fluent_ui/fluent_ui.dart';
 
-import 'base.dart';
-import 'theme.dart';
-
 class IconButton extends BaseButton {
   const IconButton({
     Key? key,