+++ lib/src/controls/surfaces/flyout/controller.dart
@@ -1,4 +1,3 @@
-import 'package:fluent_ui/fluent_ui.dart';
 import 'package:flutter/foundation.dart';
 
 class FlyoutController extends ChangeNotifier with Diagnosticable {